requirements.txt
asgiref==3.2.10
Django==3.1.1
django-environ==0.4.5
django-js-asset==1.2.2
django-mptt==0.11.0
psycopg2-binary==2.8.6
python-memcached==1.59
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
